+++ sockeye/arguments.py
@@ -104,6 +104,15 @@ def load_args(fname: str) -> argparse.Namespace:
         return argparse.Namespace(**yaml.safe_load(inp))
 
 
+class Removed(argparse.Action):
+    """
+    When this argument is specified, raise an error with the argument's help
+    message.  This is used to notify users when arguments are removed.
+    """
+    def __call__(self, parser, namespace, values, option_string=None):
+        raise RuntimeError(self.help)
+
+
 def regular_file() -> Callable:
     """
     Returns a method that can be used in argument parsing to check the argument is a regular file or a symbolic link,
@@ -468,18 +477,24 @@ def add_bucketing_args(params):
 
     params.add_argument('--bucket-width',
                         type=int_greater_or_equal(1),
-                        default=10,
+                        default=8,
                         help='Width of buckets in tokens. Default: %(default)s.')
 
-    params.add_argument('--no-bucket-scaling',
+    params.add_argument('--bucket-scaling',
                         action='store_true',
-                        help='Disable scaling source/target buckets based on length ratio. Default: %(default)s.')
+                        help='Scale source/target buckets based on length ratio to reduce padding. Default: '
+                             '%(default)s.')
+    params.add_argument('--no-bucket-scaling',
+                        action=Removed,
+                        nargs=0,
+                        help='Removed: The argument "--no-bucket-scaling" has been removed because this is now the '
+                             'default behavior. To activate bucket scaling, use the argument "--bucket-scaling".')
 
     params.add_argument(C.TRAINING_ARG_MAX_SEQ_LEN,
                         type=multiple_values(num_values=2, greater_or_equal=1),
-                        default=(99, 99),
-                        help='Maximum sequence length in tokens.'
-                             'Use "x:x" to specify separate values for src&tgt. Default: %(default)s.')
+                        default=(95, 95),
+                        help='Maximum sequence length in tokens, not counting BOS/EOS tokens (internal max sequence '
+                             'length is X+1). Use "x:x" to specify separate values for src&tgt. Default: %(default)s.')
 
 
 def add_prepare_data_cli_args(params):
@@ -720,24 +735,33 @@ def add_batch_args(params, default_batch_size=4096):
     params.add_argument('--batch-size', '-b',
                         type=int_greater_or_equal(1),
                         default=default_batch_size,
-                        help='Mini-batch size. Note that depending on the batch-type this either refers to '
-                             'words or sentences.'
-                             'Sentence: each batch contains X sentences, number of words varies. '
-                             'Word: each batch contains (approximately) X words, number of sentences varies. '
-                             'Default: %(default)s.')
-    params.add_argument("--batch-type",
+                        help='Mini-batch size per process. Depending on --batch-type, this either refers to words or '
+                             'sentences. The effective batch size (update size) is num_processes * batch_size * '
+                             'update_interval. Default: %(default)s.')
+    params.add_argument('--batch-type',
                         type=str,
                         default=C.BATCH_TYPE_WORD,
-                        choices=[C.BATCH_TYPE_SENTENCE, C.BATCH_TYPE_WORD],
-                        help="Sentence: each batch contains X sentences, number of words varies."
-                             "Word: each batch contains (approximately) X target words, "
-                             "number of sentences varies. Default: %(default)s.")
+                        choices=C.BATCH_TYPES,
+                        help='sentence: each batch contains exactly X sentences. '
+                             'word: each batch contains approximately X target words. '
+                             'max-word: each batch contains at most X target words. '
+                             'Default: %(default)s.')
+    params.add_argument('--batch-sentences-multiple-of',
+                        type=int,
+                        default=8,
+                        help='For word and max-word batching, guarantee that each batch contains a multiple of X '
+                             'sentences. For word batching, round up or down to nearest multiple. For max-word '
+                             'batching, always round down. Default: %(default)s.')
     params.add_argument('--round-batch-sizes-to-multiple-of',
+                        action=Removed,
+                        help='Removed: The argument "--round-batch-sizes-to-multiple-of" has been renamed to '
+                             '"--batch-sentences-multiple-of".')
+    params.add_argument('--update-interval',
                         type=int,
                         default=1,
-                        help='For word-based batches, round each bucket\'s batch size (measured in sentences) to a '
-                             'multiple of this integer. Default: %(default)s.')
-
+                        help='Accumulate gradients over X batches for each model update. Set a value higher than 1 to '
+                             'simulate large batches (ex: batch_size 2560 with update_interval 4 gives effective batch '
+                             'size 10240). Default: %(default)s.')
 
 def add_hybridization_arg(params):
     params.add_argument('--no-hybridization',
@@ -780,10 +804,6 @@ def add_training_args(params):
                               choices=C.METRICS,
                               help='Metric to optimize with early stopping {%(choices)s}. Default: %(default)s.')
 
-    train_params.add_argument('--update-interval',
-                              type=int,
-                              default=1,
-                              help="Number of batch gradients to accumulate before updating. Default: %(default)s.")
     train_params.add_argument(C.TRAIN_ARGS_CHECKPOINT_INTERVAL,
                               type=int_greater_or_equal(1),
                               default=4000,
@@ -872,9 +892,11 @@ def add_training_args(params):
 
     train_params.add_argument('--horovod',
                               action='store_true',
-                              help='Use Horovod/OpenMPI for distributed training (Sergeev and Del Balso 2018, '
-                                   'arxiv.org/abs/1802.05799).  When using this option, run Sockeye with `horovodrun '
-                                   '-np ... -H ... python`.')
+                              help='Use Horovod/MPI for distributed training (Sergeev and Del Balso 2018, '
+                                   'arxiv.org/abs/1802.05799). When using this option, run Sockeye with `horovodrun '
+                                   '-np X python3 -m sockeye.train` where X is the number of processes. Increasing '
+                                   'the number of processes multiplies the effective batch size (ex: batch_size 2560 '
+                                   'with `-np 4` gives effective batch size 10240).')
 
     train_params.add_argument("--kvstore",
                               type=str,