+++ test/system/test_seq_copy_sys.py
@@ -125,9 +125,9 @@ def test_seq_copy(name, train_params, translate_params, use_prepared_data, perpl
 
 @pytest.mark.parametrize(
     "name, train_params, translate_params, use_prepared_data, n_source_factors, perplexity_thresh, bleu_thresh", [
-    ("Sort:transformer:transformer",
+    ("Sort:transformer:transformer:batch_word",
      "--encoder transformer --decoder transformer"
-     " --batch-size 16 --update-interval 1 --batch-type sentence"
+     " --max-seq-len 10 --batch-size 90 --update-interval 1 --batch-type word --batch-sentences-multiple-of 1"
      " --max-updates 6000"
      " --num-layers 2 --transformer-attention-heads 2 --transformer-model-size 32 --num-embed 32"
      " --transformer-dropout-attention 0.0 --transformer-dropout-act 0.0 --transformer-dropout-prepost 0.0"
@@ -136,9 +136,9 @@ def test_seq_copy(name, train_params, translate_params, use_prepared_data, perpl
      True, 0,
      1.03,
      0.97),
-    ("Sort:transformer_with_source_factor",
+    ("Sort:transformer:transformer_with_source_factor:batch_max_word",
      "--encoder transformer --decoder transformer"
-     " --batch-size 8 --update-interval 2 --batch-type sentence"
+     " --max-seq-len 10 --batch-size 70 --update-interval 2 --batch-type max-word --batch-sentences-multiple-of 1"
      " --max-updates 6000"
      " --num-layers 2 --transformer-attention-heads 2 --transformer-model-size 32 --num-embed 32"
      " --transformer-dropout-attention 0.0 --transformer-dropout-act 0.0 --transformer-dropout-prepost 0.0"