+++ test/unit/test_data_io.py
@@ -214,13 +214,13 @@ def test_parallel_iter(source_iterables, target_iterable, expected):
 
 
 def test_sample_based_define_bucket_batch_sizes():
-    batch_by_words = False
+    batch_type = C.BATCH_TYPE_SENTENCE
     batch_size = 32
     max_seq_len = 100
     buckets = data_io.define_parallel_buckets(max_seq_len, max_seq_len, 10, 1, 1.5)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets=buckets,
                                                            batch_size=batch_size,
-                                                           batch_by_words=batch_by_words,
+                                                           batch_type=batch_type,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))
     for bbs in bucket_batch_sizes:
@@ -237,14 +237,14 @@ def test_sample_based_define_bucket_batch_sizes():
         (1.5, 1, [100.0, 50.0, 33.0, 25.0, 20.0, 20.0, 20.0, 20.0]),
         (1.5, 8, [96.0, 48.0, 32.0, 24.0, 16.0, 16.0, 16.0, 24.0])])
 def test_word_based_define_bucket_batch_sizes(length_ratio, batch_sentences_multiple_of, expected_batch_sizes):
-    batch_by_words = True
+    batch_type = C.BATCH_TYPE_WORD
     batch_num_devices = 1
     batch_size = 1000
     max_seq_len = 50
-    buckets = data_io.define_parallel_buckets(max_seq_len, max_seq_len, 10, 1, length_ratio)
+    buckets = data_io.define_parallel_buckets(max_seq_len, max_seq_len, 10, True, length_ratio)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets=buckets,
                                                            batch_size=batch_size,
-                                                           batch_by_words=batch_by_words,
+                                                           batch_type=batch_type,
                                                            batch_num_devices=batch_num_devices,
                                                            data_target_average_len=[None] * len(buckets),
                                                            batch_sentences_multiple_of=batch_sentences_multiple_of)
@@ -261,6 +261,29 @@ def test_word_based_define_bucket_batch_sizes(length_ratio, batch_sentences_mult
     assert last_bbs.batch_size >= min_expected_batch_size
     last_bbs_num_words = last_bbs.batch_size * max(*last_bbs.bucket)
     assert last_bbs_num_words >= max_num_words
+
+
+@pytest.mark.parametrize("length_ratio,batch_sentences_multiple_of,expected_batch_sizes", [
+        # Reference batch sizes manually inspected for sanity.
+        (0.5, 1, [200, 100, 66, 50, 40, 33, 28, 25, 22, 20]),
+        (1.5, 1, [100, 50, 33, 25, 20, 20, 20, 20]),
+        (1.5, 8, [96, 48, 32, 24, 16, 16, 16, 16])])
+def test_max_word_based_define_bucket_batch_sizes(length_ratio, batch_sentences_multiple_of, expected_batch_sizes):
+    batch_type = C.BATCH_TYPE_MAX_WORD
+    batch_num_devices = 1
+    batch_size = 1000
+    max_seq_len = 50
+    buckets = data_io.define_parallel_buckets(max_seq_len, max_seq_len, 10, True, length_ratio)
+    bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets=buckets,
+                                                           batch_size=batch_size,
+                                                           batch_type=batch_type,
+                                                           batch_num_devices=batch_num_devices,
+                                                           data_target_average_len=[None] * len(buckets),
+                                                           batch_sentences_multiple_of=batch_sentences_multiple_of)
+    for bbs, expected_batch_size in zip(bucket_batch_sizes, expected_batch_sizes):
+        assert bbs.batch_size == expected_batch_size
+        expected_average_target_words_per_batch = expected_batch_size * bbs.bucket[1]
+        assert bbs.average_target_words_per_batch == expected_average_target_words_per_batch
 
 
 def _get_random_bucketed_data(buckets: List[Tuple[int, int]],
@@ -311,7 +334,7 @@ def test_parallel_data_set_fill_up():
     buckets = data_io.define_parallel_buckets(100, 100, 10, 1, 1.0)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets,
                                                            batch_size,
-                                                           batch_by_words=False,
+                                                           batch_type=C.BATCH_TYPE_SENTENCE,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))
     dataset = data_io.ParallelDataSet(*_get_random_bucketed_data(buckets, min_count=1, max_count=5))
@@ -352,7 +375,7 @@ def test_parallel_data_set_permute():
     buckets = data_io.define_parallel_buckets(100, 100, 10, 1, 1.0)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets,
                                                            batch_size,
-                                                           batch_by_words=False,
+                                                           batch_type=C.BATCH_TYPE_SENTENCE,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))
     dataset = data_io.ParallelDataSet(*_get_random_bucketed_data(buckets, min_count=0, max_count=5)).fill_up(
@@ -379,7 +402,7 @@ def test_get_batch_indices():
     buckets = data_io.define_parallel_buckets(100, 100, 10, 1, 1.0)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets,
                                                            batch_size,
-                                                           batch_by_words=False,
+                                                           batch_type=C.BATCH_TYPE_SENTENCE,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))
     dataset = data_io.ParallelDataSet(*_get_random_bucketed_data(buckets=buckets,
@@ -484,7 +507,7 @@ def test_get_training_data_iters():
             target_vocab_path=None,
             shared_vocab=True,
             batch_size=batch_size,
-            batch_by_words=False,
+            batch_type=C.BATCH_TYPE_SENTENCE,
             batch_num_devices=1,
             max_seq_len_source=train_max_length,
             max_seq_len_target=train_max_length,
@@ -551,7 +574,7 @@ def test_parallel_sample_iter():
     bucket_counts = [0] + [None] * (len(buckets) - 1)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets,
                                                            batch_size,
-                                                           batch_by_words=False,
+                                                           batch_type=C.BATCH_TYPE_SENTENCE,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))
 
@@ -609,7 +632,7 @@ def test_sharded_parallel_sample_iter():
     bucket_counts = [0] + [None] * (len(buckets) - 1)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets,
                                                            batch_size,
-                                                           batch_by_words=False,
+                                                           batch_type=C.BATCH_TYPE_SENTENCE,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))
 
@@ -682,7 +705,7 @@ def test_sharded_parallel_sample_iter_num_batches():
     num_batches = num_shards * num_batches_per_shard
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets,
                                                            batch_size,
-                                                           batch_by_words=False,
+                                                           batch_type=C.BATCH_TYPE_SENTENCE,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))
 
@@ -716,7 +739,7 @@ def test_sharded_and_parallel_iter_same_num_batches():
     num_batches = num_batches_per_bucket * len(buckets)
     bucket_batch_sizes = data_io.define_bucket_batch_sizes(buckets,
                                                            batch_size,
-                                                           batch_by_words=False,
+                                                           batch_type=C.BATCH_TYPE_SENTENCE,
                                                            batch_num_devices=1,
                                                            data_target_average_len=[None] * len(buckets))